opravy chyb

+++ LARAVEL/stranka/app/helpers.php
@@ -316,33 +316,55 @@ function getYear( $date ){
 	return (int)explode(".", $date)[2];
 }
 
-function storage_deletor($type = "", $folder_hash = ''){
-	if(!is_string($type) || !is_string($folder_hash)){
+function storage_deletor($type = ""){
+	if(!is_string($type)){
 		return false;
 	}
 
+	$res = DB::table('deletor')->where('type', $type)->get();
+	if(!$res){
+		return true;
+	}
+
+	
+
 	$path = base_path('storage/app/public/');
-	$redirect = null;
-	switch($type){
-		case 'news':
-			$path .= 'news/'.$folder_hash;
-			$redirect = '/news-admin';
-		break;
-		case 'documents':
-			$path .= 'documents/'.$folder_hash;
-			$redirect = '/documents-admin';
-		break;
-		case 'subjects':
-			$path .= 'subjects/'.$folder_hash;
-			$redirect = '/subjects-admin';
-		break;
-		default:
-			return false;
+	foreach($res as $r){
+		switch($type){
+			case 'news':
+				$check = DB::table('news')->where('hash_id', $r->path)->first();
+				if($check){
+					return true;
+				}
+				$path .= 'news/'.$r->path;
+			break;
+			case 'documents':
+				$tmp = explode('/', $r->path);
+				$check = DB::table('documents')->where('hash_name', $tmp[1])->first();
+				if($check){
+					return true;
+				}
+				$path .= 'documents/'.$r->path;
+			break;
+			case 'subjects':
+				$tmp = explode('/', $r->path);
+				$check = DB::table('subjects_subcategories')->where('hash_name', $tmp[1])->first();
+				if($check){
+					return true;
+				}
+				$path .= 'subjects/'.$r->path;
+			break;
+			default:
+				return false;
+		}
+		
+		if(is_dir($path)){
+			array_map('unlink', glob("$path/*.*"));
+			rmdir($path);
+			DB::table('deletor')->where('type', $type)->where('path', $r->path)->delete();
+			return true;
+		}
 	}
-	array_map('unlink', glob("$path/*.*"));
-	if(is_dir($path)){
-		rmdir($path);
-		return redirect($redirect);
-	}
+	
 	return false;
 }
